adding script 0... again

+++ singly_linked_lists/0-print_list.c
@@ -10,13 +10,15 @@
 
 size_t print_list(const list_t *h)
 {
-	int count = 0;
+	size_t count = 0;  /* To count the number of nodes */
 
+    /* If the list is empty, return 0 */
 	if (h == NULL)
 	{
 		return (0);
 	}
 
+    /* Traverse through the list and print the data */
 	while (h != NULL)
 	{
 		if (h->str == NULL)
@@ -28,8 +30,9 @@ size_t print_list(const list_t *h)
 			printf("[%d] %s\n", h->len, h->str);
 		}
 
-	h = h->next;
-	count++;
+		h = h->next;  /* Move to the next node */
+		count++;  /* Increment the node count */
 	}
-	return (count);
+
+	return (count);  /* Return the total number of nodes */
 }
